# backend/docs/swagger.yaml
basePath: /
definitions:
  domain.Score:
    properties:
      clear_time:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      max_combo:
        type: integer
      player_name:
        type: string
      score:
        type: integer
    type: object
  httpapi.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  httpapi.RankingsResponse:
    properties:
      rankings:
        items:
          $ref: '#/definitions/domain.Score'
        type: array
    type: object
  usecase.CreateScoreInput:
    properties:
      clear_time:
        type: integer
      max_combo:
        type: integer
      player_name:
        type: string
      score:
        type: integer
    type: object
host: api.uomi.dev
info:
  contact: {}
  description: Score management API.
  title: Score API
  version: "1.0"
paths:
  /api/rankings:
    get:
      description: Returns scores ordered by score (desc), then clear time (asc),
        then created time (asc).
      operationId: getRankings
      parameters:
      - description: Max number of entries to return (1-100, default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.RankingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Get rankings
      tags:
      - scores
  /api/scores:
    post:
      consumes:
      - application/json
      description: Saves a player's score, max combo, and clear time.
      operationId: saveScore
      parameters:
      - description: Score payload
        in: body
        name: score
        required: true
        schema:
          $ref: '#/definitions/usecase.CreateScoreInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Score'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Save score
      tags:
      - scores
swagger: "2.0"